+++ NEWS.md
@@ -1,8 +1,8 @@
 # gtrendsR (development version)
 
-- Added support for NULL proxy domain, and additional curl options (#339).
+- Added support for NULL proxy domain, and additional curl options (#339) @alfirrell.
 
-- Handling related topics with only "Top" or "Rising" sections, as well as those with both.
+- Handling related topics with only "Top" or "Rising" sections, as well as those with both @alfirrell.
 
 # gtrendsR 1.4.7